Use BigDecimal for SOL amounts instead of double
Monetary values should not use floating-point arithmetic. Changes
double to BigDecimal in SolanaTransaction.amount, LargeTransfer.amount,
LargeTransferFilter threshold, and TransferRepository.findByMinAmount.
Uses string constructor (new BigDecimal("1.0")) to avoid precision loss
and compareTo() for threshold comparison.

+++ prism/src/main/java/com/stablebridge/prism/domain/model/LargeTransfer.java
@@ -1,10 +1,12 @@
 package com.stablebridge.prism.domain.model;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 
 @Builder(toBuilder = true)
 public record LargeTransfer(
         String signature,
         long slot,
-        double amount
+        BigDecimal amount
 ) {}
