+++ prism/src/main/java/com/stablebridge/prism/domain/port/TransferRepository.java
@@ -1,5 +1,6 @@
 package com.stablebridge.prism.domain.port;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.stablebridge.prism.domain.model.LargeTransfer;
@@ -8,7 +9,7 @@ public interface TransferRepository {
 
     void bulkInsert(List<LargeTransfer> transfers);
 
-    List<LargeTransfer> findByMinAmount(double minAmount, long limit, long offset);
+    List<LargeTransfer> findByMinAmount(BigDecimal minAmount, long limit, long offset);
 
-    long countByMinAmount(double minAmount);
+    long countByMinAmount(BigDecimal minAmount);
 }
